Added simple comments

+++ src/main/kotlin/domain/contract/Inventory.kt
@@ -1,5 +1,8 @@
 package domain.contract
 
+/**
+ * Интерфейс описывающий инвентарный объект
+ */
 interface Inventory {
-    val number: Int
+    val number: Long
 }
