+++ src/main/kotlin/domain/models/types/Monkey.kt
@@ -8,5 +8,5 @@ data class Monkey(
     override val health: Int,
     override val food: Int,
     override val name: String,
-    override val number: Int = nextInventoryNumber
+    override val number: Long = nextInventoryNumber
 ) : Herbivore()
