+++ src/main/kotlin/domain/models/types/Rabbit.kt
@@ -8,5 +8,5 @@ data class Rabbit(
     override val health: Int,
     override val food: Int,
     override val name: String,
-    override val number: Int = nextInventoryNumber,
+    override val number: Long = nextInventoryNumber,
 ) : Herbivore()
