+++ src/main/kotlin/domain/models/types/Wolf.kt
@@ -7,5 +7,5 @@ data class Wolf(
     override val health: Int,
     override val food: Int,
     override val name: String,
-    override val number: Int = nextInventoryNumber,
+    override val number: Long = nextInventoryNumber,
 ) : Predator()
